+++ src/config.rs
@@ -524,7 +524,7 @@ pub struct ConfigBlocklist {
     #[serde(default, deserialize_with = "ConfigBlocklist::deserialize_leaders")]
     pub leaders: HashSet<Pubkey>,
     #[serde(default, deserialize_with = "ConfigBlocklist::deserialize_leaders")]
-    pub banned_signers: HashSet<Pubkey>,
+    pub banned_accounts: HashSet<Pubkey>,
 }
 
 impl ConfigBlocklist {